+++ Sources/KeyboardKit/_Deprecated/Callouts+Curve.swift
@@ -1,5 +1,5 @@
 //
-//  CalloutCurve.swift
+//  Callouts+Curve.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-01-08.
@@ -11,8 +11,7 @@ import SwiftUI
 
 public extension Callouts {
     
-    /// This curve is used to smoothen the part of a callout,
-    /// where the callout body meets the button area.
+    @available(*, deprecated, message: "This shape has been replaced by callout-specific curve shapes.")
     struct Curve: Shape {
         
         public func path(in rect: CGRect) -> Path {
@@ -29,20 +28,4 @@ public extension Callouts {
             return path
         }
     }
-}
-
-#Preview {
-    
-    VStack(spacing: 0) {
-        RoundedRectangle(cornerRadius: 20)
-        HStack(alignment: .top, spacing: 0) {
-            Callouts.Curve().rotation3DEffect(.degrees(180), axis: (x: 0, y: 1, z: 0))
-                .frame(height: 20)
-            Color.black
-            Callouts.Curve()
-                .frame(height: 20)
-        }.frame(width: 100, height: 50)
-    }
-    .frame(height: 100)
-    .padding()
 }